latest_requirements.txt
geocoder==1.38.1
gevent==1.4.0
gTTS==2.1.1
opencv_python==4.2.0.34
librosa==0.8.0
scipy==1.2.1
playsound==1.2.2
certifi==2020.6.20
numpy==1.16.4
scikit_image==0.15.0
Werkzeug==0.15.4
tensorflow==2.3.1
keras==2.4.3
imutils==0.5.3
pandas==1.1.2
urllib3==1.24.2
requests==2.24.0
Flask==1.1.2
azure-common==1.1.25
azure-iot-device==2.1.4
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.9.3
ipython==7.16.1
Pillow==7.2.0
pyaudio==0.2.11
pyodbc==4.0.30
dlib==19.21.0
